Move getVersions to loaders
