+++ lib/reporters/allTests.js
@@ -3,7 +3,7 @@ import path from 'node:path';
 
 import { minifiers } from '../loaders/loadAllMinifiers.js';
 import { appendHistory } from '../history.js';
-import { getVersions } from '../minify.js';
+import { getVersions } from '../loaders/loadVersions.js.js';
 import { resultsAsHTML } from '../results-as-html.js';
 
 const __dirname = import.meta.dirname;
@@ -153,7 +153,7 @@ export const reportTestResults = async function (allTests, results) {
     }
   }
 
-  const versions = getVersions(minifiers);
+  const versions = getVersions();
   const history = await appendHistory(allTests, results, versions);
   await resultsAsHTML(allTests, results, versions, history, 'index.html');
   console.log('HTML report written to index.html');
